Fixed sign-compare warnings

+++ libfreerdp/crypto/crypto.c
@@ -315,7 +315,7 @@ static const char*   general_name_type_labels[] = { "OTHERNAME",
 static const char* general_name_type_label(int general_name_type)
 {
 	if ((0 <= general_name_type)
-	    && (general_name_type < ARRAYSIZE(general_name_type_labels)))
+	    && ((size_t)general_name_type < ARRAYSIZE(general_name_type_labels)))
 	{
 		return general_name_type_labels[general_name_type];
 	}
